Cleaned up app.py + PYTEST intergration
=================================================================================================== 1 error in 0.04s ====================================================================================================
(venv) shane.d@Shanes-Mac-mini neo_dolfin % pytest
================================================================================================== test session starts ==================================================================================================
platform darwin -- Python 3.11.4, pytest-7.4.0, pluggy-1.2.0
rootdir: /Users/shane.d/Documents/GitHub/Neo_Dolfin/neo_dolfin
plugins: flask-1.2.0
collected 1 item

test/test_app.py .                                                                                                                                                                                                [100%]

=================================================================================================== warnings summary ====================================================================================================
venv/lib/python3.11/site-packages/flask_wtf/recaptcha/widgets.py:2
venv/lib/python3.11/site-packages/flask_wtf/recaptcha/widgets.py:2
  /Users/shane.d/Documents/GitHub/Neo_Dolfin/neo_dolfin/venv/lib/python3.11/site-packages/flask_wtf/recaptcha/widgets.py:2: DeprecationWarning: 'flask.Markup' is deprecated and will be removed in Flask 2.4. Import 'markupsafe.Markup' instead.
    from flask import Markup

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
============================================================================================= 1 passed, 2 warnings in 0.36s =============================================================================================
(venv) shane.d@Shanes-Mac-mini neo_dolfin %
