+++ neo_dolfin/test/test_app.py
@@ -1,0 +1,9 @@
+# tests/test_app.py
+from app import app
+
+def test_hello():
+    client = app.test_client()
+
+    response = client.get('/')
+    assert response.status_code == 200
+    assert "DolFin | Landing" in response.data.decode('utf-8')
